bigdata/requirements.txt
fastapi
uvicorn
apscheduler
sqlalchemy
databases
aiomysql
pymysql
python-dotenv
hdfs
pyspark
bs4
pandas
pydantic
requests
beautifulsoup4
urllib3
python-multipart
cryptography
pytz
scikit-learn
numpy